docs: add API reference
