+++ docs/package.json
@@ -27,6 +27,7 @@
     "sharp": "^0.32.5",
     "starlight-image-zoom": "^0.9.0",
     "starlight-links-validator": "^0.13.2",
+    "starlight-openapi": "^0.8.3",
     "tailwind-merge": "^2.5.5",
     "tailwindcss": "^3.4.15",
     "tailwindcss-animate": "^1.0.7",
