+++ src/lib/server/utils/auth.ts
@@ -65,35 +65,6 @@ export const usernameExists = async (username: string) => {
   return users.length > 0;
 };
 
-export const validateApiKey = async (
-  request: Request,
-): Promise<User | null> => {
-  const apiKey = request.headers.get('Authorization')?.split('Bearer ')[1];
-  if (!apiKey) {
-    return null;
-  }
-  const hash = hashSha256(apiKey);
-  const user = await db
-    .selectFrom('user')
-    .where(
-      'id',
-      '=',
-      db.selectFrom('apiKey').where('key', '=', hash).select('userId'),
-    )
-    .selectAll()
-    .executeTakeFirst();
-
-  if (user) {
-    await db
-      .updateTable('apiKey')
-      .set({ lastUsed: new Date() })
-      .where('key', '=', hash)
-      .execute();
-  }
-
-  return user || null;
-};
-
 export const createApiKey = async (userId: string, name: string) => {
   const key = generateString();
   const hash = hashSha256(key);
